fix(apps): add circleci annotations
